pyproject.toml
[project]
name = "remainder-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "al-ov73",email = "klever1987@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "aiogram (>=3.17.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
